Bugfix for [ssue #89: https://github.com/vtourraine/AcknowList/issues/89
On iOS 15 using Xcode 13 RC, the initializer will crash unless the init(style:) initializer is overridden.

Solution taken from https://stackoverflow.com/questions/69092599/xcode-13-beta-5-error-uiviewcontroller-is-missing-its-initial-trait-collection

As I describe in the issue, the struct Acknow cannot be exposed to @objc. Therefore, a class AcknowClass was specified that is used in the initializer and then transformed into the Acknow struct. This is really hacky and breaks all projects using the init(acknowledgements:style:) initializer directly, but it solves the crash in iOS 15 when using Xcode 13 RC.